Now uses sqlite3 for database I/O.
The tab-delimited text files will still exist, but they are only for input to
a newly added script which will read them and write them to a database file.
There are currently some issues in Unicode-ASCII conversion.  Considering
moving to Python 3.x for default Unicode support.  I hope that's "a thing".
Because I certainly don't want to put 'u' in front of all the string literals!
And so far, the ways I've found on setting the default encoding are very
kludgy.