Added a basic web interface (prototype).
To test, create a "cgi-bin" directory, copy the python scripts into it, run a
web server ("python -m http.server --cgi" will do just fine), and load up
"localhost:8000/pathfinder_item_generator.html" in a browser.
Unfortunately, because scripts have to be in a scripts directory, this is a
little cumbersome, as the scripts need to be refreshed when changed.  It would
be nice to arrange the file structure to prevent the need for the copying, but
the name "cgi-bin" would not be conducive to the python module system.  So,
a different directory would perhaps be used, and the web server configured to
access that directory for cgi scripts.