Restructured files, added PF licensing info.
Moved the non-execution Python scripts to the cgi-bin directory (and the
execution python script, webgen.py).  Because they are now there, generate.py
will not work as expected.  Ideally, this application should be usable as
either a command line utility, or as a web form with a straightforward
deployment method (currently, deployment is manual, so minimal directory
"remaps" are ideal).