Corrected a problem with initdb.py.
The database "compiler", initdb.py, always left the final table empty, which
I corrected by adding a call to "commit()" after the tables were constructed.
Also implemented single item generation mode, command line only.  It is a
simple matter to implement the attachment in webgen.py, so that will be the
next step.