Further progress on implementing the treasure allocator.
Now, the Hoard page loads all treasure types when loaded, and checking boxes
simply hides the resulting list. Not sure whether a hidden list should also be
cleared of items, probably, but for now I won't do it.

Added a handler for the 0, +, and - buttons, but all it does now is print a
console log message. This seems like a good commit point.