Reorganized directory structure.
The directory structure facilitates simpler deployment from a working
development setup to my web host. Yes, "my", as although this is ostensibly an
open-source project that anyone can fork and possibly contribute and what not,
I'm basing it off of my preference for deployment, since who else is _really_
going to deal with this stuff? And besides, probably the least common
denominator, anyway. No idea if there is a better system for this sort of
thing, as with an installer for regular software, so this will have to do.